Added examples. Simple kv storage with networking.
The protocol is intended to be very simple.

The protocol itself is line oriented that is each command is delimited
with \r\n. Each line represents a command:

SET key value

Sets new value in the database for key "key" and value "value".

GET key

Gets value from storage for given key.